requirements.txt
asgiref==3.8.1
Django==5.1.4
django-appconf==1.0.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-multiselectfield==0.1.13
django-select2==8.2.3
gunicorn==23.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
sqlparse==0.5.3
